// shared/types/server/nodeDesignDto.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Omnition Api Documentation
 * Документация Api для Omnition
 * OpenAPI spec version: 1.0
 */

export interface NodeDesignDto {
    /** ID дизайна узла */
    id: number;
    /** Название вариации конструкции узлов продуктов */
    name: string;
    /** Фото дизайна узла */
    photo: string;
}